Separated wrapper from map processing functions.
